fix issues where we load metta data even if the database is set as cypher
